Move the root reducer to a better home
