+++ src/modules/ReduxTypes.js
@@ -2,8 +2,8 @@
 
 import type { Dispatch as ReduxDispatch, Store as ReduxStore } from 'redux'
 
-import type { RootState as State } from '../reducers/scenes/MainReducer.js'
-import type { Action } from './Action.js'
+import { type RootState as State } from '../reducers/RootReducer.js'
+import { type Action } from './Action.js'
 
 export type CurrencyCode = string
 export type Id = string
